Project/requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
google-generativeai>=0.3.2
openpyxl>=3.1.2
selenium>=4.15.2
python-dotenv>=1.0.0
